# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nutrition Clinic API",
    "version": "2.0.0",
    "description": "واجهة REST لنظام إدارة عيادة التغذية. مصمَّمة ليستهلكها موقع الويب وتطبيق الموبايل (Android/iOS) الخاص بالمريض دون أي تعديل. المصادقة: POST /api/auth/login ثم إرسال `Authorization: Bearer <token>` في كل طلب. التوكن صالح 12 ساعة.",
    "contact": {
      "name": "Clinic IT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "تطوير محلي"
    },
    {
      "url": "https://api.clinic.example",
      "description": "الإنتاج"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/visits": {
      "get": {
        "tags": [
          "measurements"
        ],
        "summary": "زيارات مريض",
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ items }"
          }
        }
      },
      "post": {
        "tags": [
          "measurements"
        ],
        "summary": "تسجيل زيارة",
        "responses": {
          "201": {
            "description": "Visit"
          }
        }
      }
    },
    "/api/measurements": {
      "get": {
        "tags": [
          "measurements"
        ],
        "summary": "القياسات (وزن، طول، BMI، خصر، صدر، ورك، دهون)",
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ items: Measurement[] }"
          }
        }
      },
      "post": {
        "tags": [
          "measurements"
        ],
        "summary": "تسجيل قياسات زيارة — BMI و WHR يُحسبان في السيرفر",
        "responses": {
          "201": {
            "description": "Measurement"
          }
        }
      }
    },
...
    "/api/appointments": {
      "get": {
        "tags": [
          "appointments"
        ],
        "summary": "قائمة مواعيد مع فلاتر",
        "parameters": [
          {
            "name": "patient_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "scheduled",
                "confirmed",
                "done",
                "cancelled",
                "no_show"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ items }"
          }
        }
      },
      "post": {
        "tags": [
          "appointments"
        ],
        "summary": "تسجيل موعد (يرفض التداخل 409)",
        "responses": {
          "201": {
            "description": "Appointment"
          },
          "409": {
            "description": "الوقت محجوز"
          }
        }
      }
    },
    "/api/appointments/today": {
      "get": {
        "tags": [
          "appointments"
        ],
        "summary": "مواعيد اليوم + ملخص",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TodaySchedule"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}